Fixing folder names and code consistency
Folder names are now uncapitalised and the code is more consistent. I use tidyverse for a lot of things instead of different packages that can do the same thing. Also I fixed the issue of us not being consistent with loading datasets, and fitting models. Further, a lot of things did not run, I fixed those problems.